- Minor bugs fixed - Added pagination to controllers
