--- src/main/java/by/modsen/pizza/dto/CategoryDto.java
@@ -1,9 +1,0 @@
-package by.modsen.pizza.dto;
-
-import lombok.*;
-
-@Data
-public class CategoryDto {
-    private String name;
-    private String description;
-}
