+++ src/main/java/com/modsen/pizza/dto/CategoryDto.java
@@ -1,0 +1,7 @@
+package com.modsen.pizza.dto;
+
+import lombok.*;
+
+@Data
+public class CategoryDto {
+}
