+++ src/main/java/com/modsen/pizza/entity/OrderItem.java
@@ -1,4 +1,4 @@
-package by.modsen.pizza.entity;
+package com.modsen.pizza.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -29,5 +29,5 @@ public class OrderItem {
     private Integer quantity;
 
     @Column(name = "price")
-    private double price;
+    private Double price;
 }
