+++ src/main/java/com/modsen/pizza/entity/User.java
@@ -1,4 +1,4 @@
-package by.modsen.pizza.entity;
+package com.modsen.pizza.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
