+++ src/main/java/com/modsen/pizza/service/OrderService.java
@@ -1,11 +1,12 @@
-package by.modsen.pizza.service;
+package com.modsen.pizza.service;
 
-import by.modsen.pizza.dto.OrderDto;
+import com.modsen.pizza.dto.OrderDto;
+import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 public interface OrderService {
     OrderDto createOrder(OrderDto orderDto);
     OrderDto getOrderById(Long id);
-    List<OrderDto> getAllOrders();
+    List<OrderDto> getAllOrders(Pageable pageable);
     List<OrderDto> getOrdersByUserId(Long userId);
 }
